requirements.py3.txt
Django==2.0.8
django-wysiwyg-redactor==0.5.1
